requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
discord==2.3.2
discord.py==2.3.2
dnspython==2.6.1
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
pymongo==4.6.3
python-dotenv==1.0.1
yarl==1.9.4
